fix: rename Audience to If
