+++ packages/ui/app/src/mdx/components/index.tsx
@@ -4,7 +4,6 @@ import dynamic from "next/dynamic";
 import { ComponentProps, PropsWithChildren, ReactElement } from "react";
 import { FernErrorBoundary, FernErrorBoundaryProps, FernErrorTag } from "../../components/FernErrorBoundary";
 import { AccordionGroup } from "./accordion";
-import { Audience } from "./audience";
 import { Availability } from "./availability";
 import { Badge } from "./badge";
 import { Bleed } from "./bleed";
@@ -28,6 +27,7 @@ import { Column, ColumnGroup } from "./columns";
 import { Frame } from "./frame";
 import { A, HeadingRenderer, Image, Li, Ol, Strong, Ul } from "./html";
 import { Table } from "./html-table";
+import { If } from "./if";
 import { IFrame } from "./iframe";
 import { Mermaid } from "./mermaid";
 import { ParamField } from "./mintlify";
@@ -41,7 +41,6 @@ const LaunchDarkly = dynamic(() => import("./launchdarkly/LaunchDarkly").then((m
 
 const FERN_COMPONENTS = {
     AccordionGroup,
-    Audience,
     Availability,
     Badge,
     Bleed,
@@ -59,6 +58,7 @@ const FERN_COMPONENTS = {
     EndpointResponseSnippet,
     Frame,
     Icon: RemoteFontAwesomeIcon,
+    If,
     LaunchDarkly,
     Mermaid,
     ParamField,
