+++ packages/ui/fern-docs-utils/src/constants.ts
@@ -9,6 +9,11 @@ export const HEADER_X_FERN_HOST = "x-fern-host" as const;
 export const HEADER_X_MATCHED_PATH = "x-matched-path" as const;
 
 /**
+ * The role that is used to represent everyone (including unauthenticated users)
+ */
+export const EVERYONE_ROLE = "everyone" as const;
+
+/**
  * Revalidate all cached docs within 6 days because S3 signatures expire after 7 days.
  */
 export const REVALIDATE_SECONDS = 60 * 60 * 24 * 6;
